chore: unify context files
The first implementation was better. This fixes #1302.
