+++ packages/react-day-picker/src/contexts/SelectMultiple/SelectMultipleContext.tsx
@@ -1,10 +1,14 @@
 import React from 'react';
-import { isDayPickerMultiple } from '../../types/DayPickerMultiple';
-import { DayPickerProps } from '../../types/DayPicker';
 
-import { DayClickEventHandler } from '../../types/EventHandlers';
-import { Modifiers } from '../../types/Modifiers';
-import { SelectMultipleProviderInternal } from './SelectMultipleProviderInternal';
+import { isSameDay } from 'date-fns';
+
+import {
+  DayClickEventHandler,
+  DayPickerMultipleProps,
+  DayPickerProps,
+  isDayPickerMultiple,
+  Modifiers
+} from '../../types';
 
 /** Represent the modifiers that are changed by the multiple selection. */
 export type SelectMultipleModifiers = Pick<Modifiers, 'selected' | 'disabled'>;
@@ -57,5 +61,79 @@ export function SelectMultipleProvider(
       initialProps={props.initialProps}
       children={props.children}
     />
+  );
+}
+
+type SelectMultipleProviderInternalProps = {
+  initialProps: DayPickerMultipleProps;
+  children: React.ReactNode;
+};
+
+export function SelectMultipleProviderInternal({
+  initialProps,
+  children
+}: SelectMultipleProviderInternalProps): JSX.Element {
+  const { selected } = initialProps;
+  const handleDayClick: DayClickEventHandler = (day, modifiers, e) => {
+    initialProps.onDayClick?.(day, modifiers, e);
+
+    const isMinSelected = Boolean(
+      initialProps.min &&
+        modifiers.selected &&
+        selected &&
+        selected.length === initialProps.min
+    );
+    if (isMinSelected) {
+      return;
+    }
+    const isMaxSelected = Boolean(
+      initialProps.max &&
+        !modifiers.selected &&
+        selected &&
+        selected.length === initialProps.max
+    );
+    if (isMaxSelected) {
+      return;
+    }
+
+    const days = selected ? [...selected] : [];
+    if (modifiers.selected) {
+      const index = days.findIndex((selectedDay) =>
+        isSameDay(day, selectedDay)
+      );
+      days.splice(index, 1);
+    } else {
+      days.push(day);
+    }
+    initialProps.onSelect?.(days, day, modifiers, e);
+  };
+
+  const modifiers: SelectMultipleModifiers = {
+    selected: [],
+    disabled: []
+  };
+
+  if (selected) {
+    modifiers.selected = selected;
+    modifiers.disabled = [
+      function disableDay(day: Date) {
+        const isMaxSelected =
+          initialProps.max &&
+          selected &&
+          selected.length > initialProps.max - 1;
+        const isSelected = selected?.some((selectedDay) =>
+          isSameDay(selectedDay, day)
+        );
+        return Boolean(isMaxSelected && !isSelected);
+      }
+    ];
+  }
+
+  const contextValue = { selected, handleDayClick, modifiers };
+
+  return (
+    <SelectMultipleContext.Provider value={contextValue}>
+      {children}
+    </SelectMultipleContext.Provider>
   );
 }
